+++ tests/message.rs
@@ -48,7 +48,7 @@ fn main() {
                     let (_handles, body, sessions) =
                         tpm_unmarshal_command(&original_bytes).unwrap();
 
-                    let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE];
+                    let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE as usize];
                     let built_len = {
                         let mut writer = TpmWriter::new(&mut built_bytes);
                         let tag = if sessions.is_empty() {
@@ -76,7 +76,7 @@ fn main() {
                             .expect("unmarshaling failed on a success test case")
                             .expect("expected success but got TpmRc error");
 
-                        let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE];
+                        let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE as usize];
                         let built_len = {
                             let mut writer = TpmWriter::new(&mut built_bytes);
                             let rc = TpmRc::from(TpmRcBase::Success);
@@ -100,7 +100,7 @@ fn main() {
 
                         assert_eq!(actual_rc.base(), expected_rc_base, "Mismatched TpmRc error");
 
-                        let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE];
+                        let mut built_bytes = [0u8; TPM_MAX_COMMAND_SIZE as usize];
                         let built_len = {
                             let mut writer = TpmWriter::new(&mut built_bytes);
                             tpm_marshal_response(
